Made default filter using Canvas
