+++ app/src/main/java/com/android/wardrobeManager/ui/images/filters/HorizontalStripes.java
@@ -1,13 +1,23 @@
-package com.android.wardrobeManager.ui.images;
+package com.android.wardrobeManager.ui.images.filters;
 
 import android.graphics.Bitmap;
 
+import com.android.wardrobeManager.ui.images.DesignFilterManager;
+
 import java.util.Map;
 
-public class ThinVerticalStripes implements DesignFilterManager.DesignFilter {
+public class HorizontalStripes implements DesignFilterManager.DesignFilter {
 
     protected Map<String, DesignFilterManager.DesignFilter> filters;
-    private int thickness = 10;
+    private int thickness;
+
+    public HorizontalStripes(int thickness) {
+        this.thickness = thickness;
+    }
+
+    public HorizontalStripes() {
+        thickness = 10;
+    }
 
     public Bitmap filter(Bitmap bitmap, Bitmap ref, int[] colors) {
 
@@ -17,7 +27,7 @@ public class ThinVerticalStripes implements DesignFilterManager.DesignFilter {
             int y = i / bitmap.getWidth();
 
             if (ref.getPixel(x, y) != 0xFFFFFFFF) {
-                int stripeColor = colors[(x / thickness) % (colors.length)];
+                int stripeColor = colors[(y / thickness) % (colors.length)];
 
                 bitmap.setPixel(i % bitmap.getWidth(), i / bitmap.getWidth(), stripeColor);
